Prestige Elm Park Whitefield Bangalore
Prestige Elm Park Whitefield Bangalore is a premium development that offers the best urban park accommodation in Silicon Valley. This vast land area is spread over wide acres where about 86% is devoted to open and green spaces.

It is a large space group housing project that welcomes you to the ancient world with extraordinary privileges. Every unit is a masterpiece from where you can clearly see the reserved greens of Bengaluru and an exotic villa community. So, a beautiful life awaits you in this great project which is not just a blessing.

Prestige Elm Park Bangalore is located in a prime location best of both worlds which blends the luxuries of city life with the freshness of nature. It is an elite address that is loved by everyone who wants to live a luxurious life.

Master planning is done to ensure that all the high-rise towers are in place for uninterrupted and extraordinary views. Providing easy access to all world-class amenities and facilities for the unparalleled pleasure of staying in these upscale apartments. It is big on amenities, lifestyle, and elegance, here you dream big and live even bigger. It comes with rare features to kick off your day in a classy atmosphere. This project offers amenities like a Swimming pool, tennis court, Spa, Gym, cricket ground, professional pitch, green outfield, pavilion sit-out, and two exclusive clubhouses with indoor games to make life better.

Prestige Elm Park floor plans are offering 1 BHK, 2 BHK, and 3 BHK residential apartments that come in various sizes. Enjoy everything from the comfort of your lavish living destination that is spacious and expertly designed just for you.

For More Details
Visit us - https://bit.ly/3wdy3uA